Added migratsiya

+++ RoboticsPos/Common/DTOs/ProductDTO.cs
@@ -5,12 +5,12 @@ public class ProductDTO : BaseDTO
     public string Name { get; set; }
     public string BarCode { get; set; }
     
-    public int  AmountInPackage { get; set; }
+    public decimal  AmountInPackage { get; set; }
    // public int CompanyId { get; set; }
-    public int Amount { get; set; }
-    public int ActualPrice { get; set; }
-    public int Price { get; set; }
-    public int PriceOfPiece { get; set; }
+    public decimal Amount { get; set; }
+    public decimal ActualPrice { get; set; }
+    public decimal Price { get; set; }
+    public decimal PriceOfPiece { get; set; }
     public bool Selected { get; set; } = false;
     public long? DiscountId { get; set; }
 }
